+++ java/believe/mob/StationaryEnemy.kt
@@ -17,5 +17,5 @@ internal class StationaryEnemy(
 
     override fun unbind() {}
 
-    override fun update(delta: Int) {}
+    override fun update(delta: Long) {}
 }
